getting job data from session-db
